Add a partition-aware version of new order.
This will be slightly less efficient than the older version for "local" new order transactions, but it avoids code duplication.
NOTE: This assumes that s_data and s_dist_xx columns of stock are fully replicated. I don't think this is a good idea, but this is what the previous work did.